+++ lib/pages/game/game_add.dart
@@ -1,0 +1,39 @@
+import 'package:flutter/material.dart';
+import 'package:game_list/db/model/game.dart';
+
+class GameAdd extends StatelessWidget {
+  static const routeName = '/gameAdd';
+
+  final _formKey = GlobalKey<FormState>();
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Add game'),
+      ),
+      floatingActionButton: FloatingActionButton(
+        child: Icon(Icons.check),
+        onPressed: () {
+          if (_formKey.currentState.validate()) {
+            Navigator.pop(context);
+          }
+        },
+      ),
+      body: Padding(
+          padding: const EdgeInsets.all(8.0),
+          child: Form(
+              key: _formKey,
+              child: Column(
+                children: [
+                  TextFormField(
+                    decoration: InputDecoration(labelText: 'Name'),
+                  ),
+                  TextFormField(
+                    decoration: InputDecoration(labelText: 'Cover URL'),
+                  )
+                ],
+              ))),
+    );
+  }
+}